Added contact pages

+++ components/ScrollSection/ScrollSection.tsx
@@ -6,35 +6,39 @@ import Text from "../Text/Text";
 
 function ScrollSection() {
   const sectionRef = useRef(null);
+  const slideRef = useRef(null);
   const triggerRef = useRef(null);
 
   gsap.registerPlugin(ScrollTrigger);
 
   useEffect(() => {
-    const pin = gsap.fromTo(
-      sectionRef.current,
-      {
-        translateX: 0,
-      },
-      {
-        translateX: "-300vw",
-        ease: "none",
-        duration: 1,
-        scrollTrigger: {
-          trigger: triggerRef.current,
-          start: "top top",
-          end: "2000 top",
-          scrub: 0.6,
-          pin: true,
+    if (sectionRef && sectionRef.current) {
+      const pin = gsap.fromTo(
+        sectionRef.current,
+        {
+          translateX: 0,
         },
-      }
-    );
-    return () => {
-      {
-        /* A return function for killing the animation on component unmount */
-      }
-      pin.kill();
-    };
+        {
+          translateX: "-300vw",
+          ease: "none",
+          duration: 1,
+          scrollTrigger: {
+            trigger: triggerRef.current,
+            start: "top top",
+            end: "2000 top",
+            scrub: 0.6,
+            pin: true,
+          },
+        }
+      );
+      return () => {
+        {
+          /* A return function for killing the animation on component unmount */
+        }
+        pin.kill();
+      };
+      return;
+    }
   }, []);
 
   return (
@@ -57,6 +61,7 @@ function ScrollSection() {
             ></div>
             <div className="relative flex items-center  h-screen w-1/2 p-8">
               <Text>
+                <h1 className="text-6xl py-2 text-blue font-header">مرحباً</h1>
                 <h1 className="text-9xl py-2 font-header">
                   Welcome to Baba's Pantry
                 </h1>
@@ -96,9 +101,28 @@ function ScrollSection() {
               </Text>
             </div>
           </div>
+
           {/* Section 3 */}
           <div className="scroll-section bg-purple">
-            <h3>Section 3</h3>
+            <div className="relative items-center flex h-screen w-1/2 p-8">
+              <Text>
+                <h1 className="text-6xl py-2 font-header">قائمة طعام</h1>
+                <h1 className="text-9xl py-2 font-header">Menu</h1>
+                <p className="text-lg">
+                  Hungry for a delicious feast? Click on our menu photo and
+                  explore a world of mouth-watering flavors in a new page! From
+                  savory starters to indulgent desserts, satisfy your cravings
+                  with just a click.
+                </p>
+              </Text>
+            </div>
+            <div className="relative items-center flex h-screen w-1/2 p-8">
+              <img src="/images/MENU-FRONT.png"></img>
+            </div>
+
+            <div className="relative items-center flex h-screen w-1/2 p-8">
+              <img src="/images/MENU-BACK.png"></img>
+            </div>
           </div>
           {/* Section */}
           <div className="scroll-section bg-green">
